# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CLARK Gateway",
    "version": "3.7.6",
    "description": "The API gateway for all of the CLARK microservices",
    "license": { "name": "ISC", "url": "https://www.isc.org/licenses/" },
    "contact": {
      "name": "SecurEd Inc.",
      "url": "https://secured.team/",
      "email": "info@secured.team"
    },
    "x-logo": {
      "url": "https://clark.center/assets/images/logo.png",
      "altText": "CLARK Logo"
    },
    "x-tagGroups": [
      { "name": "Library Service", "tags": ["Library Service"] },
      { "name": "Notification Service", "tags": ["Notification Service"] },
      { "name": "Feature Service", "tags": ["Feature Service"] },
      { "name": "Rating Service", "tags": ["Rating Service"] },
      { "name": "Outcome Service", "tags": ["Outcome Service"] },
      {
        "name": "Learning Object Service",
        "tags": ["Learning Object Service"]
      },
      { "name": "User Service", "tags": ["User Service"] },
      { "name": "Utility Service", "tags": ["Utility Service"] },
      {
        "name": "Standard Guideline Service",
        "tags": ["Standard Guideline Service"]
      }
    ]
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "Development" },
    { "url": "https://api-gateway.clark.center", "description": "Production" }
  ],
  "paths": {
...
    "/users/{username}/learning-objects/{id}/relevancy-check": {
      "patch": {
        "description": "Updates the nextCheck attribute",
        "tags": ["Learning Object Service"],
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "schema": { "type": "string" },
            "required": true,
            "description": "The learning object's author's username"
          },
          {
            "in": "path",
            "name": "id",
            "schema": { "type": "string" },
            "required": true,
            "description": "The object id of the learning object"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "year": {
                    "type": "number",
                    "required": true,
                    "description": "The amount of years from now in the future (1, 2, or 3)",
                    "example": 3
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "OK" },
          "400": {
            "description": "BAD REQUEST - The year in the body is missing or invalid"
          },
          "401": { "description": "UNAUTHENTICATED - User not logged in" },
          "403": {
            "description": "UNAUTHORIZED - User does not have editor or admin privileges"
          },
          "404": { "description": "NOT FOUND - Learning object not found" }
        }
      }
    },
...
    "/users/{username}/learning-objects/{id}/learning-outcomes/{outcomeId}/guidelines": {
      "patch": {
        "description": "Updates the guidelines attribute",
        "tags": ["Learning Object Service"],
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "schema": { "type": "string" },
            "required": true,
            "description": "The learning object's author's username"
          },
          {
            "in": "path",
            "name": "id",
            "schema": { "type": "string" },
            "required": true,
            "description": "The object id of the learning object"
          },
          {
            "in": "path",
            "name": "outcomeId",
            "schema": { "type": "string" },
            "required": true,
            "description": "The id of the outcome being updated"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "guidelines": {
                    "type": "array",
                    "required": true,
                    "description": "The complete list of guidelines that should belong to an outcome at any given time",
                    "example": ["000000000000000000000000"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "OK" },
          "400": {
            "description": "BAD REQUEST - The guidelines body is empty, or the ids are invalid mongo ids"
          },
          "401": { "description": "UNAUTHENTICATED - User not logged in" },
          "403": {
            "description": "UNAUTHORIZED - User does not have editor, admin, or curator privileges"
          },
          "404": {
            "description": "NOT FOUND - Learning object not found, Guidelines not found, Outcome not found"
          }
        }
      }
    },
    "/users/{username}/learning-objects/{id}/topics": {
      "patch": {
        "description": "Updates topics on a object",
        "tags": ["Learning Object Service"],
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "schema": { "type": "string" },
            "required": true,
            "description": "The learning object's author's username"
          },
          {
            "in": "path",
            "name": "id",
            "schema": { "type": "string" },
            "required": true,
            "description": "The object id of the learning object"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "topicIds": {
                    "type": "array",
                    "items": { "type": "string" },
                    "required": true,
                    "description": "An array of topic ids"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "OK" },
          "400": {
            "description": "BAD REQUEST - Topic ids provided is not a valid id"
          },
          "401": { "description": "UNAUTHENTICATED - User not logged in" },
          "403": {
            "description": "UNAUTHORIZED - User does not have editor or admin privileges"
          },
          "404": {
            "description": "NOT FOUND - Learning object or topic not found"
          }
        }
      }
    },
    "/users/{username}/learning-objects/{id}/topics/{topicId}": {
      "delete": {
        "description": "Deletes a topic from a object",
        "tags": ["Learning Object Service"],
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "schema": { "type": "string" },
            "required": true,
            "description": "The learning object's author's username"
          },
          {
            "in": "path",
            "name": "id",
            "schema": { "type": "string" },
            "required": true,
            "description": "The object id of the learning object"
          },
          {
            "in": "path",
            "name": "topicId",
            "schema": { "type": "string" },
            "required": true,
            "description": "The id of the topic"
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "400": {
            "description": "BAD REQUEST - Topic id provided is not a valid id"
          },
          "401": { "description": "UNAUTHENTICATED - User not logged in" },
          "403": {
            "description": "UNAUTHORIZED - User does not have editor or admin privileges"
          },
          "404": {
            "description": "NOT FOUND - Learning object or topic not found"
          }
        }
      }
    },
    "/learning-objects/evaluators": {
      "post": {
        "description": "Assigns evaluators to objects",
        "tags": ["Learning Object Service"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cuids": {
                    "type": "array",
                    "items": { "type": "string" },
                    "required": true,
                    "description": "An array of cuids"
                  },
                  "assignerIds": {
                    "type": "array",
                    "items": { "type": "string" },
                    "required": true,
                    "description": "An array of user ids"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "OK" },
          "400": {
            "description": "BAD REQUEST - Assigner id(s) provided do not have an access group"
          },
          "401": { "description": "UNAUTHENTICATED - User not logged in" },
          "403": {
            "description": "UNAUTHORIZED - User does not have curator, editor, or admin privileges"
          },
          "404": {
            "description": "NOT FOUND - Learning objects or users were not found"
          },
          "409": {
            "description": "CONFLICT - A given user has already been assigned to one or more cuids"
          }
        }
      },
      "patch": {
        "description": "Unassigns evaluators to objects",
        "tags": ["Learning Object Service"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cuids": {
                    "type": "array",
                    "items": { "type": "string" },
                    "required": true,
                    "description": "An array of cuids"
                  },
                  "assignerIds": {
                    "type": "array",
                    "items": { "type": "string" },
                    "required": true,
                    "description": "An array of user ids"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": { "description": "NO CONTENT" },
          "401": { "description": "UNAUTHENTICATED - User not logged in" },
          "403": {
            "description": "UNAUTHORIZED - User does not have curator, editor, or admin privileges"
          },
          "404": {
            "description": "NOT FOUND - Learning objects or users were not found"
          }
        }
      }
    },
    "/users/{username}/learning-objects/{cuid}/evaluation": {
      "patch": {
        "description": "Evaluates a learning object",
        "tags": ["Learning Object Service"],
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "schema": { "type": "string" },
            "required": true,
            "description": "The learning object's author's username"
          },
          {
            "in": "path",
            "name": "cuid",
            "schema": { "type": "string" },
            "required": true,
            "description": "The cuid of the learning object"
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "401": { "description": "UNAUTHENTICATED - User not logged in" },
          "403": {
            "description": "UNAUTHORIZED - User does not have mapper, reviewer, curator, editor, or admin privileges"
          },
          "404": { "description": "NOT FOUND - Learning object was not found" }
        }
      }
    },
    "/users/:username/evaluations?status": {
      "get": {
        "description": "Gets evaluator's assigned learning objects",
        "tags": ["Learning Object Service"],
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "schema": { "type": "string" },
            "required": true,
            "description": "The evaluator's username"
          },
          {
            "in": "query",
            "name": "status",
            "schema": { "type": "string" },
            "required": false,
            "description": "The requested status of evaluated learning objects"
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "401": { "description": "UNAUTHENTICATED - User not logged in" },
          "403": {
            "description": "UNAUTHORIZED - User does not have curator, editor, or admin privileges"
          },
          "404": {
            "description": "NOT FOUND - Learning objects or users were not found"
          }
        }
      }
    },
...
  },
...
}